+++ packages/web/lib/pages/snapinmanagementpage.class.php
@@ -170,14 +170,14 @@ class SnapinManagementPage extends FOGPage {
                 ->set('description',$_REQUEST['description'])
                 ->set('file',$snapinfile)
                 ->set('args',$_REQUEST['args'])
-                ->set('reboot',(int)(isset($_REQUEST['action']) && $_REQUEST['action'] === 'reboot'))
-                ->set('shutdown',(string)intval((int)(isset($_REQUEST['action']) && $_REQUEST['action'] === 'shutdown')))
+                ->set('reboot',(isset($_REQUEST['action']) && $_REQUEST['action'] === 'reboot'))
+                ->set('shutdown',(string)intval((isset($_REQUEST['action']) && $_REQUEST['action'] === 'shutdown')))
                 ->set('runWith',$_REQUEST['rw'])
                 ->set('runWithArgs',$_REQUEST['rwa'])
-                ->set('isEnabled',(string)intval((int)isset($_REQUEST['isEnabled'])))
-                ->set('toReplicate',(string)intval((int)isset($_REQUEST['toReplicate'])))
+                ->set('isEnabled',(string)intval(isset($_REQUEST['isEnabled'])))
+                ->set('toReplicate',(string)intval(isset($_REQUEST['toReplicate'])))
                 ->set('hide',(string)intval(isset($_REQUEST['isHidden'])))
-                ->set('timeout',(int)$_REQUEST['timeout'])
+                ->set('timeout',$_REQUEST['timeout'])
                 ->addGroup($_REQUEST['storagegroup']);
             if (!$Snapin->save()) throw new Exception(_('Add snapin failed!'));
             self::$HookManager->processEvent('SNAPIN_ADD_SUCCESS',array('Snapin'=>&$Snapin));
@@ -346,15 +346,15 @@ class SnapinManagementPage extends FOGPage {
                     ->set('description',$_REQUEST['description'])
                     ->set('file',$snapinfile)
                     ->set('args',$_REQUEST['args'])
-                    ->set('reboot',(int)(isset($_REQUEST['action']) && $_REQUEST['action'] === 'reboot'))
-                    ->set('shutdown',(string)intval((int)(isset($_REQUEST['action']) && $_REQUEST['action'] === 'shutdown')))
+                    ->set('reboot',(isset($_REQUEST['action']) && $_REQUEST['action'] === 'reboot'))
+                    ->set('shutdown',(string)intval((isset($_REQUEST['action']) && $_REQUEST['action'] === 'shutdown')))
                     ->set('runWith',$_REQUEST['rw'])
                     ->set('runWithArgs',$_REQUEST['rwa'])
-                    ->set('protected',(int)isset($_REQUEST['protected_snapin']))
-                    ->set('isEnabled',(string)intval((int)isset($_REQUEST['isEnabled'])))
-                    ->set('toReplicate',(string)intval((int)isset($_REQUEST['toReplicate'])))
+                    ->set('protected',isset($_REQUEST['protected_snapin']))
+                    ->set('isEnabled',(string)intval(isset($_REQUEST['isEnabled'])))
+                    ->set('toReplicate',(string)intval(isset($_REQUEST['toReplicate'])))
                     ->set('hide',(string)intval(isset($_REQUEST['isHidden'])))
-                    ->set('timeout',(int)$_REQUEST['timeout']);
+                    ->set('timeout',$_REQUEST['timeout']);
                 break;
             case 'snap-storage':
                 $this->obj->addGroup($_REQUEST['storagegroup']);
