+++ packages/web/lib/reg-task/taskqueue.class.php
@@ -10,11 +10,11 @@ class TaskQueue extends TaskingElement {
                     if (!$this->Task->save()) throw new Exception(_('Failed to update task'));
                     $MulticastSession = self::getClass('MulticastSessions',@max(self::getSubObjectIDs('MulticastSessionsAssociation',array('taskID'=>$this->Task->get('id')),'msID')));
                     if (!$MulticastSession->isValid()) throw new Exception(_('Invalid Session'));
-                    $clientCount = (int)$MulticastSession->get('clients');
+                    $clientCount = $MulticastSession->get('clients');
                     $MulticastSession->set('clients',++$clientCount);
                     if (!$MulticastSession->save()) throw new Exception(_('Failed to update session'));
-                    $CheckedInCount = (int)self::getClass('MulticastSessionsAssociationManager')->count(array('msID'=>$MulticastSession->get('id')));
-                    $sessionClientCount = (int)$MulticastSession->get('sessclients');
+                    $CheckedInCount = self::getClass('MulticastSessionsAssociationManager')->count(array('msID'=>$MulticastSession->get('id')));
+                    $sessionClientCount = $MulticastSession->get('sessclients');
                     $SessionStateID = ($CheckedInCount == $clientCount || ($sessionClientCount > 0 && $clientCount > 0)) ? 3 : 1;
                     $this->Task->set('stateID',$this->getProgressState());
                     if (!$this->Task->save()) throw new Exception(_('Failed to update task'));
@@ -43,7 +43,7 @@ class TaskQueue extends TaskingElement {
                         foreach ($this->StorageNodes AS &$StorageNode) {
                             if (!$StorageNode->isValid()) continue;
                             if ($StorageNode->get('maxClients') < 1) continue;
-                            $nodeAvailableSlots = (int)$StorageNode->get('maxClients') - (int)$StorageNode->getUsedSlotCount();
+                            $nodeAvailableSlots = $StorageNode->get('maxClients') - $StorageNode->getUsedSlotCount();
                             if ($nodeAvailableSlots < 1) continue;
                             if (!isset($tmpStorageNode)) {
                                 $tmpStorageNode = self::nodeFail($StorageNode,$this->Host->get('id'));
